pyproject.toml
[project]
name = "convexobstaclesvcc"
version = "0.1.0"
description = "Visibility clique cover (VCC) for approximating free space in Euclidean domains with convex obstacles."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "drake>=1.30.0",
    "matplotlib>=3.7",
    "networkx>=3.0",
    "numba>=0.65.0",
    "numpy>=1.24",
    "scipy>=1.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/convexobstaclesvcc"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "ruff>=0.6",
]
